requirements.txt
django==3.0.2
psycopg2==2.8.4
django-background-tasks==1.2.5
pika==1.1.0
django-environ==0.4.5
